web-eq-server/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic_settings
pydantic
email-validator
SQLAlchemy>=1.4.16
psycopg2-binary>=2.8.6
python-multipart
python-jose[cryptography]>=3.3.0
bcrypt==4.0.1
passlib>=1.7.4
httpx
aiohttp

